+++ scheduling/src/main/java/com/nabla/sdk/scheduling/scene/ScheduleAppointmentActivity.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.FragmentTransaction.TRANSIT_FRAGMENT_MATCH_ACTIVITY
 import androidx.fragment.app.FragmentTransaction.TRANSIT_FRAGMENT_OPEN
 import androidx.fragment.app.commit
 import com.benasher44.uuid.Uuid
+import com.nabla.sdk.core.annotation.NablaInternal
 import com.nabla.sdk.core.ui.helpers.requireSdkName
 import com.nabla.sdk.core.ui.helpers.setSdkName
 import com.nabla.sdk.core.ui.helpers.viewBinding
@@ -18,6 +19,7 @@ import com.nabla.sdk.scheduling.domain.entity.CategoryId
 import com.nabla.sdk.scheduling.scene.slots.TimeSlotsFragment
 import kotlinx.datetime.Instant
 
+@NablaInternal
 public class ScheduleAppointmentActivity : AppCompatActivity() {
 
     private val binding by viewBinding(NablaSchedulingActivityScheduleAppointmentHostBinding::inflate)
@@ -60,8 +62,8 @@ public class ScheduleAppointmentActivity : AppCompatActivity() {
         }
     }
 
-    public companion object {
-        public fun newIntent(context: Context, name: String): Intent = Intent(context, ScheduleAppointmentActivity::class.java).apply {
+    internal companion object {
+        fun newIntent(context: Context, name: String): Intent = Intent(context, ScheduleAppointmentActivity::class.java).apply {
             setSdkName(name)
         }
     }